backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pinyin==0.4.0
cohere==4.37
openai==0.28.1
Werkzeug==2.3.7
gunicorn==21.2.0
flask-cors==4.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.0
flask-sqlalchemy==3.0.5
